refactor: attachment service 패키지 생성
