+++ src/main/java/techcourse/fakebook/service/attachment/assembler/AttachmentAssembler.java
@@ -1,9 +1,9 @@
-package techcourse.fakebook.service.article.assembler;
+package techcourse.fakebook.service.attachment.assembler;
 
 import org.springframework.stereotype.Component;
 import techcourse.fakebook.domain.article.ArticleAttachment;
 import techcourse.fakebook.domain.user.UserProfileImage;
-import techcourse.fakebook.service.article.dto.AttachmentResponse;
+import techcourse.fakebook.service.attachment.dto.AttachmentResponse;
 
 @Component
 public class AttachmentAssembler {
