+++ src/main/java/techcourse/fakebook/service/user/LoginService.java
@@ -8,7 +8,7 @@ import techcourse.fakebook.domain.user.User;
 import techcourse.fakebook.domain.user.UserRepository;
 import techcourse.fakebook.exception.NotFoundUserException;
 import techcourse.fakebook.exception.NotMatchPasswordException;
-import techcourse.fakebook.service.article.AttachmentService;
+import techcourse.fakebook.service.attachment.AttachmentService;
 import techcourse.fakebook.service.user.dto.LoginRequest;
 import techcourse.fakebook.service.user.dto.UserOutline;
 import techcourse.fakebook.service.user.encryptor.Encryptor;
