+++ src/test/java/techcourse/fakebook/service/ServiceTestHelper.java
@@ -3,15 +3,12 @@ package techcourse.fakebook.service;
 import io.restassured.internal.util.IOUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.web.multipart.MultipartFile;
 import techcourse.fakebook.domain.user.UserProfileImage;
-import techcourse.fakebook.service.article.AttachmentService;
-import techcourse.fakebook.service.article.dto.AttachmentResponse;
+import techcourse.fakebook.service.attachment.dto.AttachmentResponse;
 import techcourse.fakebook.service.user.dto.UserOutline;
-import techcourse.fakebook.service.user.dto.UserSignupRequest;
 
 import java.io.File;
 import java.io.FileInputStream;
