+++ src/test/java/techcourse/fakebook/service/attachment/AttachmentServiceTest.java
@@ -1,4 +1,4 @@
-package techcourse.fakebook.service.article;
+package techcourse.fakebook.service.attachment;
 
 import io.restassured.internal.util.IOUtils;
 import org.junit.jupiter.api.Test;
@@ -9,8 +9,8 @@ import org.springframework.web.multipart.MultipartFile;
 import techcourse.fakebook.domain.article.Article;
 import techcourse.fakebook.domain.article.ArticleRepository;
 import techcourse.fakebook.exception.NotFoundArticleException;
-import techcourse.fakebook.service.article.AttachmentService;
-import techcourse.fakebook.service.article.dto.AttachmentResponse;
+import techcourse.fakebook.service.attachment.AttachmentService;
+import techcourse.fakebook.service.attachment.dto.AttachmentResponse;
 
 import java.io.File;
 import java.io.FileInputStream;
