+++ src/test/java/techcourse/fakebook/service/user/UserServiceTest.java
@@ -10,8 +10,7 @@ import techcourse.fakebook.domain.user.UserProfileImage;
 import techcourse.fakebook.domain.user.UserRepository;
 import techcourse.fakebook.exception.NotFoundUserException;
 import techcourse.fakebook.service.ServiceTestHelper;
-import techcourse.fakebook.service.article.AttachmentService;
-import techcourse.fakebook.service.article.dto.AttachmentResponse;
+import techcourse.fakebook.service.attachment.AttachmentService;
 import techcourse.fakebook.service.user.assembler.UserAssembler;
 import techcourse.fakebook.service.user.dto.*;
 import techcourse.fakebook.service.user.encryptor.Encryptor;
